# Improve the perf when resolving code lens in large java file