--- a/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/BaseFrameworkSearcher.java
+++ b/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/BaseFrameworkSearcher.java
@@ -20,6 +20,9 @@ import com.microsoft.java.test.plugin.util.TestItemUtils;
 import org.eclipse.jdt.core.IMethod;
 import org.eclipse.jdt.core.IType;
 import org.eclipse.jdt.core.JavaModelException;
+import org.eclipse.jdt.core.dom.IAnnotationBinding;
+import org.eclipse.jdt.core.dom.IMethodBinding;
+import org.eclipse.jdt.core.dom.ITypeBinding;
 import org.eclipse.jdt.core.search.IJavaSearchConstants;
 import org.eclipse.jdt.core.search.SearchPattern;
 
@@ -75,6 +78,17 @@ public abstract class BaseFrameworkSearcher implements TestFrameworkSearcher {
     }
 
     @Override
+    public boolean annotates(IAnnotationBinding[] annotations, String annotationName) {
+        for (final IAnnotationBinding annotation : annotations) {
+            final ITypeBinding annotationType = annotation.getAnnotationType();
+            if (annotationType != null && (annotationType.getQualifiedName().equals(annotationName))) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
     public TestItem parseTestItem(IMethod method) throws JavaModelException {
         return TestItemUtils.constructTestItem(method, TestLevel.METHOD, this.getTestKind());
     }
@@ -82,5 +96,11 @@ public abstract class BaseFrameworkSearcher implements TestFrameworkSearcher {
     @Override
     public TestItem parseTestItem(IType type) throws JavaModelException {
         return TestItemUtils.constructTestItem(type, TestLevel.CLASS, this.getTestKind());
+    }
+
+    @Override
+    public TestItem parseTestItem(IMethodBinding methodBinding) throws JavaModelException {
+        return TestItemUtils.constructTestItem((IMethod) methodBinding.getJavaElement(),
+            TestLevel.METHOD, this.getTestKind());
     }
 }