--- a/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/JUnit5TestSearcher.java
+++ b/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/JUnit5TestSearcher.java
@@ -205,7 +205,7 @@ public class JUnit5TestSearcher extends BaseFrameworkSearcher {
             final ITypeBinding annotationType = annotationBinding.getAnnotationType();
             if (annotationType != null && hierarchy.add(annotationType)) {
                 if (matchesName(annotationType, annotationName) ||
-                        matchesNameInAnnotationHierarchy(annotation, annotationName, hierarchy)) {
+                        matchesNameInAnnotationHierarchy(annotationBinding, annotationName, hierarchy)) {
                     return true;
                 }
             }