--- a/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/TestFrameworkSearcher.java
+++ b/java-extension/com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/TestFrameworkSearcher.java
@@ -40,7 +40,7 @@ public interface TestFrameworkSearcher {
 
     SearchPattern getSearchPattern();
 
-    boolean annotates(IAnnotationBinding[] annotations, String annotationName);
+    boolean findAnnotation(IAnnotationBinding[] annotations, String[] annotationNames);
 
     @Deprecated
     TestItem parseTestItem(IMethod method) throws JavaModelException;