--- a/test/test-projects/junit5/src/test/java/junit5/FastTest.java
+++ b/test/test-projects/junit5/src/test/java/junit5/FastTest.java
@@ -1,0 +1,16 @@
+package junit5;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+@Tag("fast")
+@Test
+public @interface FastTest {
+}