--- a/test/test-projects/junit5/src/test/java/junit5/MetaAnnotationTest.java
+++ b/test/test-projects/junit5/src/test/java/junit5/MetaAnnotationTest.java
@@ -1,0 +1,7 @@
+package junit5;
+
+public class MetaAnnotationTest {
+    @FastTest
+    void myFastTest() {
+    }
+}