… .vscode/launch.json
… java-extension/pom.xml
… test/shared.ts