fix #1039 

Using the attached source file in #1039, below are some data points with this PR change:

### On Xeon E5-1620 v4 @3.50GHz + 32G mem + Windows
#### Open workspace (ms)
| | 1 | 2 | 3 | 4 | 5 | AVG |
|---|---|---|---|---|---|---|
| New | 1432 | 2262 | 1036 | 2186 | 1214 | 1626 |
| Old | 8257 | 8380 | 13953 | 8678 | 10782 | 10010 |

#### Document Change - Input `Enter` (ms)
| | 1 | 2 | 3 | 4 | 5 | AVG |
|---|---|---|---|---|---|---|
| New | 380 | 284 | 278 | 258 | 274 | 295 |
| Old | 2963 | 3383 | 3122 | 2659 | 2635 | 2952 |

### On i5 @ 2.9GHz + 8G mem + MacOS
#### Open workspace (ms)
| | 1 | 2 | 3 | 4 | 5 | AVG |
|---|---|---|---|---|---|---|
| New | 2110 | 3438 | 1966 | 2387 | 2375 | 2455 |
| Old | 10714 | 8551 | 9146 | 9828 | 11441 | 9936 |

#### Document Change - Input `Enter` (ms)
| | 1 | 2 | 3 | 4 | 5 | AVG |
|---|---|---|---|---|---|---|
| New | 1760 | 395 | 345 | 436 | 405 | 668 |
| Old | 4597 | 4041 | 3863 | 3586 | 3584 | 3934 |

### Follow-ups

Besides this PR change, we can improve the perf further, including:

- Stop resolving the @DisplayName at compile time. This should be resolved at runtime actually
- Stop resolving the parameters for JUnit 5 cases, only resolve them when necessary at launching
- Skip some searchers according to the project's test dependencies
- Use text match